# Optimize Serializer.ReadFrom